Update: Documentation
Update: The documentation has been updated to fit the new release.

In this release the UI has been changed; the menu in no longer a topbar menu but a side panel menu. The generic options will still be available in the topbar, but the options that are in the main purpose category will be moved to the side panel menu.

Version has been updated to: 1.1.0-alpha.